move interpolate3d out of header
